[libra-trace] Refactor libra-trace from debug_interface.
