+++ Cargo.toml
@@ -135,6 +135,7 @@ members = [
 #
 # For more, see the "Conditional compilation for tests" section in documentation/coding_guidelines.md.
 default-members = [
+    "common/trace",
     "config/config-builder",
     "config/generate-key",
     "config/management",
