+++ common/debug-interface/Cargo.toml
@@ -12,12 +12,8 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0.31"
 tokio = { version = "0.2.21", features = ["full"] }
-serde_json = "1.0.56"
 reqwest = { version = "0.10.6", features = ["blocking", "json"], default_features = false }
-serde = "1.0.114"
-once_cell = "1.4.0"
 warp = "0.2.3"
-chrono = "0.4.13"
 
 libra-logger = { path = "../logger", version = "0.1.0" }
 libra-metrics = { path = "../metrics", version = "0.1.0" }
