+++ common/trace/Cargo.toml
@@ -14,7 +14,11 @@ chrono = "0.4.13"
 structopt = "0.3.15"
 tokio = { version = "0.2.21", features = ["full"] }
 serde_json = "1.0.56"
+serde = "1.0.114"
+anyhow = "1.0.31"
+reqwest = { version = "0.10.6", features = ["blocking", "json"], default_features = false }
+once_cell = "1.4.0"
 
-debug-interface = { path = "../../common/debug-interface", version = "0.1.0"}
 libra-logger = { path = "../../common/logger", version = "0.1.0" }
+libra-metrics = { path = "../metrics", version = "0.1.0" }
 libra-workspace-hack = { path = "../workspace-hack", version = "0.1.0" }
