+++ common/trace/src/counters.rs
