+++ consensus/src/persistent_liveness_storage.rs
@@ -6,11 +6,11 @@ use anyhow::{format_err, Context, Result};
 use consensus_types::{
     block::Block, quorum_cert::QuorumCert, timeout_certificate::TimeoutCertificate, vote::Vote,
 };
-use debug_interface::prelude::*;
 use executor_types::ExecutedTrees;
 use libra_config::config::NodeConfig;
 use libra_crypto::HashValue;
 use libra_logger::prelude::*;
+use libra_trace::prelude::*;
 use libra_types::{
     block_info::Round, epoch_change::EpochChangeProof, ledger_info::LedgerInfo,
     transaction::Version,
