+++ execution/executor/Cargo.toml
@@ -16,7 +16,6 @@ once_cell = "1.4.0"
 serde_json = "1.0.56"
 
 consensus-types = { path = "../../consensus/consensus-types", version = "0.1.0"}
-debug-interface = { path = "../../common/debug-interface", version = "0.1.0" }
 executor-types = { path = "../executor-types", version = "0.1.0" }
 lcs = { path = "../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
@@ -24,6 +23,7 @@ libra-logger = { path = "../../common/logger", version = "0.1.0" }
 libra-metrics = { path = "../../common/metrics", version = "0.1.0" }
 libra-secure-net = { path = "../../secure/net", version = "0.1.0" }
 libra-state-view = { path = "../../storage/state-view", version = "0.1.0" }
+libra-trace = { path = "../../common/trace", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
 move-core-types = { path = "../../language/move-core/types", version = "0.1.0" }
 libra-vm = { path = "../../language/libra-vm", version = "0.1.0" }
