+++ language/libra-vm/Cargo.toml
@@ -20,13 +20,13 @@ libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
 libra-logger = { path = "../../common/logger", version = "0.1.0" }
 libra-metrics = { path = "../../common/metrics", version = "0.1.0" }
 libra-state-view = { path = "../../storage/state-view", version = "0.1.0" }
+libra-trace = { path = "../../common/trace", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 move-core-types = { path = "../move-core/types", version = "0.1.0" }
 move-vm-runtime = { path = "../move-vm/runtime", version = "0.1.0" }
 move-vm-types = { path = "../move-vm/types", version = "0.1.0" }
 vm = { path = "../vm", version = "0.1.0" }
-debug-interface = { path = "../../common/debug-interface", version = "0.1.0" }
 serde_json = "1.0.56"
 serde = { version = "1.0.114", default-features = false }
