+++ testsuite/Cargo.toml
@@ -35,6 +35,7 @@ libra-secure-time = { path = "../secure/time", version = "0.1.0" }
 libra-secure-storage = { path = "../secure/storage", version = "0.1.0", features = ["testing"] }
 libra-swarm = { path = "libra-swarm", version = "0.1.0"}
 libra-temppath = { path = "../common/temppath", version = "0.1.0" }
+libra-trace = { path = "../common/trace", version = "0.1.0" }
 libra-transaction-scripts = { path = "../secure/transaction-scripts", version = "0.1.0" }
 libra-types = { path = "../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../common/workspace-hack", version = "0.1.0" }
