+++ testsuite/cluster-test/src/health/log_tail.rs
@@ -4,8 +4,7 @@
 #![forbid(unsafe_code)]
 
 use crate::{health::ValidatorEvent, util::unix_timestamp_now};
-use debug_interface::json_log::JsonLogEntry as DebugInterfaceEvent;
-use libra_logger::*;
+use libra_logger::{json_log::JsonLogEntry as DebugInterfaceEvent, *};
 use std::{
     sync::{
         atomic::{AtomicBool, AtomicI64, Ordering},
