+++ x.toml
@@ -68,7 +68,6 @@ members = [
     "common/libradoc",
     "common/proptest-helpers",
     "common/retrier",
-    "common/trace",
     "common/workspace-builder",
     "devtools/x",
     "devtools/x-core",
